Fix linter
