+++ components/core/ui/navigation/src/main/java/com/flipperdevices/core/ui/navigation/FeatureEntry.kt
@@ -1,5 +1,4 @@
 package com.flipperdevices.core.ui.navigation
 
 // https://proandroiddev.com/navigating-through-multi-module-jetpack-compose-applications-6c9a31fa12b6
-interface FeatureEntry {
-}
+interface FeatureEntry
