+++ components/filemanager/impl/src/main/java/com/flipperdevices/filemanager/impl/composable/ComposableFileManagerScreen.kt
@@ -3,19 +3,13 @@ package com.flipperdevices.filemanager.impl.composable
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
 import androidx.compose.material.Scaffold
-import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import com.flipperdevices.deeplink.api.DeepLinkParser
 import com.flipperdevices.deeplink.model.DeeplinkContent
-import com.flipperdevices.filemanager.impl.R
-import com.flipperdevices.filemanager.impl.composable.bar.ComposableEllipsizeStartText
+import com.flipperdevices.filemanager.impl.composable.bar.ComposableFileManagerTopBar
 import com.flipperdevices.filemanager.impl.composable.list.ComposableFileManagerContent
 import com.flipperdevices.filemanager.impl.model.FileItem
 import com.flipperdevices.filemanager.impl.model.FileManagerState
@@ -57,28 +51,3 @@ fun ComposableFileManagerScreen(
         )
     }
 }
-
-@Composable
-private fun ComposableFileManagerTopBar(path: String, onClickSaveButton: () -> Unit) {
-    TopAppBar(
-        title = {
-            ComposableEllipsizeStartText(
-                text = path
-            )
-        },
-        actions = {
-            if (isAbleToSafe(path)) {
-                IconButton(onClick = onClickSaveButton) {
-                    Icon(
-                        painter = painterResource(
-                            com.flipperdevices.core.ui.res.R.drawable.ic_upload
-                        ),
-                        contentDescription = stringResource(R.string.filemanager_upload_action)
-                    )
-                }
-            }
-        }
-    )
-}
-
-private fun isAbleToSafe(path: String) = path.startsWith("/ext")
