+++ components/filemanager/impl/src/main/java/com/flipperdevices/filemanager/impl/composable/bar/ComposableFileManagerTopBar.kt
@@ -1,0 +1,37 @@
+package com.flipperdevices.filemanager.impl.composable.bar
+
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.material.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import com.flipperdevices.core.ui.res.R as DesignSystem
+import com.flipperdevices.filemanager.impl.R
+
+@Composable
+fun ComposableFileManagerTopBar(path: String, onClickSaveButton: () -> Unit) {
+    TopAppBar(
+        title = {
+            ComposableEllipsizeStartText(
+                text = path
+            )
+        },
+        actions = {
+            if (isAbleToSafe(path)) {
+                IconButton(onClick = onClickSaveButton) {
+                    Icon(
+                        painter = painterResource(
+                            DesignSystem.drawable.ic_upload
+                        ),
+                        contentDescription = stringResource(
+                            R.string.filemanager_upload_action
+                        )
+                    )
+                }
+            }
+        }
+    )
+}
+
+private fun isAbleToSafe(path: String) = path.startsWith("/ext")
