--- components/filemanager/receive/.gitignore
@@ -1,1 +1,0 @@
-/build
