--- components/filemanager/receive/src/main/AndroidManifest.xml
@@ -1,5 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-        package="com.flipperdevices.filemanager.receive">
-
-    <uses-permission android:name="android.permission.READ_USER_DICTIONARY" />
-</manifest>
