--- components/filemanager/receive/src/main/java/com/flipperdevices/filemanager/receive/composable/ComposableReceive.kt
@@ -1,14 +1,0 @@
-package com.flipperdevices.filemanager.receive.composable
-
-import androidx.compose.runtime.Composable
-import com.flipperdevices.deeplink.model.DeeplinkContent
-
-@Composable
-fun ComposableReceive(
-    deeplinkContent: DeeplinkContent,
-    flipperPath: String,
-    onSuccessful: () -> Unit,
-    onCancel: () -> Unit
-
-) {
-}
