--- components/filemanager/receive/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="receive_dialog_title">Upload file %1$s</string>
-</resources>
